Time: 63 ms (6.76%), Space: 88.1 MB (97.17%) - LeetHub

+++ 2226-maximum-candies-allocated-to-k-children/2226-maximum-candies-allocated-to-k-children.cpp
@@ -6,7 +6,7 @@ public:
         int ans = 0;
         while (left <= right) {
             int mid = (left + right) / 2;
-            int cnt = 0;
+            long long cnt = 0;
             for (int i = 0; i < candies.size(); i++) {
                 cnt += candies[i] / mid;
             }
